Changed some things to be nullable. Will change more as they come up. Might be best to make it all nullable? Not sure yet.
Changed layout of the app. I was encountering issues with padding changes and things due to showing/hiding components. What i've done is added a new screen and I am navigating between the two but making it look the same. This simplifies a lot of things where I can conditionally render different bits.